Arrumando cadastrar Atendimento

+++ Clinica/src/main/java/com/fatec/Clinica/model/Atendimento.java
@@ -1,18 +1,11 @@
 package com.fatec.Clinica.model;
 
 import java.io.Serializable;
-import java.time.LocalDate;
-
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -38,7 +31,7 @@ public class Atendimento implements Serializable{
 	private float valor;
 	
 	@Column(name="id_cliente")
-    private int cliente;
+    private long cliente;
 	
 	@Column(name="id_servicos_oferecidos")
     private int servicosOferecido;
